refactor(menu): split menu subsystems
