+++ src/sf33rd/Source/Game/menu/options_menu.c
@@ -1,0 +1,929 @@
+/**
+ * @file menu.c
+ * Menus
+ */
+
+#include "sf33rd/Source/Game/menu/menu.h"
+#include "common.h"
+#include "core/app.h"
+#include "main.h"
+#include "platform/app/sdl/sdl_app.h"
+#include "sf33rd/AcrSDK/common/pad.h"
+#include "sf33rd/Source/Game/animation/appear.h"
+#include "sf33rd/Source/Game/effect/eff04.h"
+#include "sf33rd/Source/Game/effect/eff10.h"
+#include "sf33rd/Source/Game/effect/eff18.h"
+#include "sf33rd/Source/Game/effect/eff23.h"
+#include "sf33rd/Source/Game/effect/eff38.h"
+#include "sf33rd/Source/Game/effect/eff39.h"
+#include "sf33rd/Source/Game/effect/eff40.h"
+#include "sf33rd/Source/Game/effect/eff43.h"
+#include "sf33rd/Source/Game/effect/eff45.h"
+#include "sf33rd/Source/Game/effect/eff51.h"
+#include "sf33rd/Source/Game/effect/eff57.h"
+#include "sf33rd/Source/Game/effect/eff58.h"
+#include "sf33rd/Source/Game/effect/eff61.h"
+#include "sf33rd/Source/Game/effect/eff63.h"
+#include "sf33rd/Source/Game/effect/eff64.h"
+#include "sf33rd/Source/Game/effect/eff66.h"
+#include "sf33rd/Source/Game/effect/eff75.h"
+#include "sf33rd/Source/Game/effect/eff91.h"
+#include "sf33rd/Source/Game/effect/effa0.h"
+#include "sf33rd/Source/Game/effect/effa3.h"
+#include "sf33rd/Source/Game/effect/effa8.h"
+#include "sf33rd/Source/Game/effect/effc4.h"
+#include "sf33rd/Source/Game/effect/effect.h"
+#include "sf33rd/Source/Game/effect/effk6.h"
+#include "sf33rd/Source/Game/engine/grade.h"
+#include "sf33rd/Source/Game/engine/plcnt.h"
+#include "sf33rd/Source/Game/engine/pls02.h"
+#include "sf33rd/Source/Game/engine/workuser.h"
+#include "sf33rd/Source/Game/game.h"
+#include "sf33rd/Source/Game/io/gd3rd.h"
+#include "sf33rd/Source/Game/io/pulpul.h"
+#include "sf33rd/Source/Game/io/vm_sub.h"
+#include "sf33rd/Source/Game/menu/dir_data.h"
+#include "sf33rd/Source/Game/menu/ex_data.h"
+#include "sf33rd/Source/Game/message/en/msgtable_en.h"
+#include "sf33rd/Source/Game/rendering/color3rd.h"
+#include "sf33rd/Source/Game/rendering/mmtmcnt.h"
+#include "sf33rd/Source/Game/rendering/texgroup.h"
+#include "sf33rd/Source/Game/screen/entry.h"
+#include "sf33rd/Source/Game/sound/se.h"
+#include "sf33rd/Source/Game/sound/sound3rd.h"
+#include "sf33rd/Source/Game/stage/bg.h"
+#include "sf33rd/Source/Game/stage/bg_data.h"
+#include "sf33rd/Source/Game/stage/bg_sub.h"
+#include "sf33rd/Source/Game/system/pause.h"
+#include "sf33rd/Source/Game/system/ramcnt.h"
+#include "sf33rd/Source/Game/system/reset.h"
+#include "sf33rd/Source/Game/system/saver.h"
+#include "sf33rd/Source/Game/system/sys_sub.h"
+#include "sf33rd/Source/Game/system/sys_sub2.h"
+#include "sf33rd/Source/Game/system/sysdir.h"
+#include "sf33rd/Source/Game/system/work_sys.h"
+#include "sf33rd/Source/Game/ui/count.h"
+#include "sf33rd/Source/Game/ui/sc_sub.h"
+#include "sf33rd/Source/PS2/mc/savesub.h"
+#include "structs.h"
+
+#if NETPLAY_ENABLED
+#include "platform/netplay/fistbump.h"
+#include "platform/netplay/netplay.h"
+#include "port/sdl/netplay_screen.h"
+#include "sf33rd/Source/Game/menu/netplay_menu.h"
+#endif
+
+void Default_Training_Option();
+void Dummy_Move_Sub(struct _TASK* task_ptr, s16 PL_id, s16 id, s16 type, s16 max);
+void Return_Pause_Sub(struct _TASK* task_ptr);
+void Dummy_Move_Sub_LR(u16 sw, s16 id, s16 type, s16 cursor_id);
+void Return_VS_Result_Sub(struct _TASK* task_ptr);
+void Exit_Replay_Save(struct _TASK* task_ptr);
+void Setup_NTr_Data(s16 ix);
+s32 Check_Pad_in_Pause(struct _TASK* task_ptr);
+void Next_Be_Tr_Menu(struct _TASK* task_ptr);
+void Yes_No_Cursor_Exit_Training(struct _TASK* task_ptr, s16 cursor_id);
+void Check_Skip_Recording();
+void Check_Skip_Replay(s16 ix);
+void Setup_Tr_Pause(struct _TASK* task_ptr);
+void Control_Player_Tr();
+s32 Pause_Check_Tr(s16 PL_id);
+void Setup_Win_Lose_OBJ();
+s32 Pause_in_Normal_Tr(struct _TASK* task_ptr);
+void Training_Disp_Sub(struct _TASK* task_ptr);
+
+// forward decls
+void After_Title(struct _TASK* task_ptr);
+void In_Game(struct _TASK* task_ptr);
+void Wait_Load_Save(struct _TASK* task_ptr);
+void Wait_Replay_Check(struct _TASK* task_ptr);
+void Suspend_Menu();
+void Wait_Replay_Load();
+void Training_Menu(struct _TASK* task_ptr);
+void After_Replay(struct _TASK* task_ptr);
+void Wait_Pause_in_Tr(struct _TASK* task_ptr);
+void Reset_Training(struct _TASK* task_ptr);
+void Reset_Replay(struct _TASK* task_ptr);
+void End_Replay_Menu(struct _TASK* task_ptr);
+void Mode_Select(struct _TASK* task_ptr);
+void Option_Select(struct _TASK* task_ptr);
+void Training_Mode(struct _TASK* task_ptr);
+void System_Direction(struct _TASK* task_ptr);
+void Load_Replay(struct _TASK* task_ptr);
+void toSelectGame(struct _TASK* task_ptr);
+void Game_Option(struct _TASK* task_ptr);
+void Button_Config(struct _TASK* task_ptr);
+void Screen_Adjust(struct _TASK* task_ptr);
+void Sound_Test(struct _TASK* task_ptr);
+void Extra_Option(struct _TASK* task_ptr);
+void VS_Result(struct _TASK* task_ptr);
+void Save_Replay(struct _TASK* task_ptr);
+void Direction_Menu(struct _TASK* task_ptr);
+void Setup_VS_Mode(struct _TASK* task_ptr);
+void Setup_Next_Page(struct _TASK* task_ptr, u8 /* unused */);
+void Load_Replay_Sub(struct _TASK* task_ptr);
+void Button_Exit_Check(struct _TASK* task_ptr, s16 PL_id);
+void Back_to_Mode_Select(struct _TASK* task_ptr);
+void Flash_1P_or_2P(struct _TASK* task_ptr);
+
+void bg_etc_write_ex(s16 type);
+void Decide_PL(s16 PL_id);
+void imgSelectGameButton();
+void jmpRebootProgram();
+s32 Check_Pause_Term_Tr(s16 PL_id);
+
+void Menu_in_Sub(struct _TASK* task_ptr);
+s32 Exit_Sub(struct _TASK* task_ptr, s16 cursor_ix, s16 next_routine);
+u16 MC_Move_Sub(u16 sw, s16 cursor_id, s16 menu_max, s16 cansel_menu);
+s32 Menu_Sub_case1(struct _TASK* task_ptr);
+void System_Dir_Move_Sub(s16 PL_id);
+void System_Dir_Move_Sub_LR(u16 sw, s16 cursor_id);
+u16 Dir_Move_Sub2(u16 sw);
+void Dir_Move_Sub_LR(u16 sw, s16 /* unused */);
+void Ex_Move_Sub_LR(u16 sw, s16 PL_id);
+u16 Game_Option_Sub(s16 PL_id);
+u16 GO_Move_Sub_LR(u16 sw, s16 cursor_id);
+void Button_Config_Sub(s16 PL_id);
+void Button_Move_Sub_LR(u16 sw, s16 cursor_id);
+void Return_Option_Mode_Sub(struct _TASK* task_ptr);
+void Screen_Adjust_Sub(s16 PL_id);
+void Screen_Exit_Check(struct _TASK* task_ptr, s16 PL_id);
+void Screen_Move_Sub_LR(u16 sw);
+u16 Sound_Cursor_Sub(s16 PL_id);
+u16 SD_Move_Sub_LR(u16 sw);
+u16 After_VS_Move_Sub(u16 sw, s16 cursor_id, s16 menu_max);
+s32 VS_Result_Move_Sub(struct _TASK* task_ptr, s16 PL_id);
+void Training_Init(struct _TASK* task_ptr);
+void Menu_Select(struct _TASK* task_ptr);
+void Button_Config_in_Game(struct _TASK* task_ptr);
+void Character_Change(struct _TASK* task_ptr);
+void Pad_Come_Out(struct _TASK* task_ptr);
+void Normal_Training(struct _TASK* task_ptr);
+void Blocking_Training(struct _TASK* task_ptr);
+void Dummy_Setting(struct _TASK* task_ptr);
+void Training_Option(struct _TASK* task_ptr);
+void Button_Config_Tr(struct _TASK* task_ptr);
+void Blocking_Tr_Option(struct _TASK* task_ptr);
+void Training_Init_Sub(struct _TASK* task_ptr);
+void Training_Exit_Sub(struct _TASK* task_ptr);
+void Menu_Init(struct _TASK* task_ptr);
+s32 Check_Pad_in_Pause(struct _TASK* task_ptr);
+s32 Yes_No_Cursor_Move_Sub(struct _TASK* task_ptr);
+void Setup_Button_Sub(s16 x, s16 y, s16 master_player);
+void Button_Exit_Check_in_Game(struct _TASK* task_ptr, s16 PL_id);
+void Setup_Save_Replay_1st(struct _TASK* task_ptr);
+s32 Save_Replay_MC_Sub(struct _TASK* task_ptr, s16 /* unused */);
+void Button_Exit_Check_in_Tr(struct _TASK* task_ptr, s16 PL_id);
+s32 VS_Result_Select_Sub(struct _TASK* task_ptr, s16 PL_id);
+void Setup_Replay_Sub(s16 type, MenuHeader char_type, s16 master_player);
+extern const u8 Setup_Index_64[];
+
+void Game_Option(struct _TASK* task_ptr) {
+    s16 char_index;
+    s16 ix;
+
+    s16 unused_s3;
+    s16 unused_s2;
+
+    switch (task_ptr->r_no[2]) {
+    case 0:
+        FadeOut(1, 0xFF, 8);
+        task_ptr->r_no[2] += 1;
+        task_ptr->timer = 5;
+        Menu_Common_Init();
+        Menu_Cursor_Y[0] = 0;
+        Menu_Suicide[1] = 1;
+        Menu_Suicide[2] = 0;
+        Menu_Cursor_Y[0] = 0;
+        Menu_Cursor_Y[1] = 0;
+        Order[0x4F] = 4;
+        Order_Timer[0x4F] = 1;
+        Order[0x4E] = 2;
+        Order_Dir[0x4E] = 2;
+        Order_Timer[0x4E] = 1;
+        effect_57_init(0x6A, MENU_HEADER_GAME_OPTION, 0, 0x3F, 2);
+        Order[0x6A] = 1;
+        Order_Dir[0x6A] = 8;
+        Order_Timer[0x6A] = 1;
+
+        for (ix = 0, unused_s3 = char_index = 0x19; ix < 0xC; ix++, unused_s2 = char_index++) {
+            effect_61_init(0, ix + 0x50, 0, 2, char_index, ix, 0x70A7);
+            Order[ix + 0x50] = 1;
+            Order_Dir[ix + 0x50] = 4;
+            Order_Timer[ix + 0x50] = ix + 0x14;
+        }
+
+        Menu_Cursor_Move = 0xA;
+
+        for (ix = 0; ix < 0xA; ix++) {
+            effect_64_init(ix + 0x5D, 0, 2, Setup_Index_64[ix], ix, 0x70A7, ix + 1, 0, 0);
+            Order[ix + 0x5D] = 1;
+            Order_Dir[ix + 0x5D] = 4;
+            Order_Timer[ix + 0x5D] = ix + 0x14;
+        }
+
+        break;
+
+    case 1:
+        Menu_Sub_case1(task_ptr);
+        break;
+
+    case 2:
+        if (FadeIn(1, 0x19, 8) != 0) {
+            task_ptr->r_no[2] += 1;
+            Suicide[3] = 0;
+        }
+
+        break;
+
+    case 3:
+        Game_Option_Sub(0);
+        Button_Exit_Check(task_ptr, 0);
+        Game_Option_Sub(1);
+        Button_Exit_Check(task_ptr, 1);
+        Save_Game_Data();
+        break;
+
+    default:
+        Exit_Sub(task_ptr, 2, 5);
+        break;
+    }
+}
+
+u16 Game_Option_Sub(s16 PL_id) {
+    u16 sw;
+    u16 ret;
+
+    sw = ~plsw_01[PL_id] & plsw_00[PL_id];
+    sw = Check_Menu_Lever(PL_id, 0);
+    ret = MC_Move_Sub(sw, 0, 0xB, 0xFF);
+    ret |= GO_Move_Sub_LR(sw, 0);
+    ret &= 0x20F;
+    return ret;
+}
+
+const u8 Game_Option_Index_Data[10] = { 7, 3, 3, 3, 3, 1, 1, 1, 1, 1 };
+
+u16 GO_Move_Sub_LR(u16 sw, s16 cursor_id) {
+    if (Menu_Cursor_Y[cursor_id] > 9) {
+        return 0;
+    }
+
+    switch (sw) {
+    case 4:
+        Convert_Buff[0][cursor_id][Menu_Cursor_Y[cursor_id]] -= 1;
+
+        if (Convert_Buff[0][cursor_id][Menu_Cursor_Y[cursor_id]] < 0) {
+            Convert_Buff[0][cursor_id][Menu_Cursor_Y[cursor_id]] = Game_Option_Index_Data[Menu_Cursor_Y[cursor_id]];
+        }
+
+        SE_dir_cursor_move();
+        return 4;
+
+    case 8:
+        Convert_Buff[0][cursor_id][Menu_Cursor_Y[cursor_id]] += 1;
+
+        if (Convert_Buff[0][cursor_id][Menu_Cursor_Y[cursor_id]] > Game_Option_Index_Data[Menu_Cursor_Y[cursor_id]]) {
+            Convert_Buff[0][cursor_id][Menu_Cursor_Y[cursor_id]] = 0;
+        }
+
+        SE_dir_cursor_move();
+        return 8;
+
+    default:
+        return 0;
+    }
+}
+
+void Button_Config(struct _TASK* task_ptr) {
+    s16 ix;
+    s16 disp_index;
+
+    switch (task_ptr->r_no[2]) {
+    case 0:
+        FadeOut(1, 0xFF, 8);
+        task_ptr->r_no[2] += 1;
+        task_ptr->timer = 5;
+        Menu_Common_Init();
+        pp_operator_check_flag(0);
+        Menu_Cursor_Y[0] = 0;
+        Menu_Cursor_Y[1] = 0;
+        Menu_Suicide[1] = 1;
+        Menu_Suicide[2] = 0;
+        Copy_Key_Disp_Work();
+        Order[0x4F] = 4;
+        Order_Timer[0x4F] = 1;
+        Order[0x4E] = 2;
+        Order_Dir[0x4E] = 2;
+        Order_Timer[0x4E] = 1;
+        effect_57_init(0x6B, MENU_HEADER_BUTTON_CONFIG, 0, 0x3F, 2);
+        Order[0x6B] = 1;
+        Order_Dir[0x6B] = 8;
+        Order_Timer[0x6B] = 1;
+
+        for (ix = 0; ix < 12; ix++) {
+            effect_23_init(0, ix + 0x50, 0, 2, 2, ix, 0x70A7, ix + 9, 1);
+            Order[ix + 0x50] = 1;
+            Order_Dir[ix + 0x50] = 4;
+            Order_Timer[ix + 0x50] = ix + 0x14;
+            effect_23_init(1, ix + 0x5C, 0, 2, 3, ix, 0x70A7, ix + 9, 1);
+            Order[ix + 0x5C] = 1;
+            Order_Dir[ix + 0x5C] = 4;
+            Order_Timer[ix + 0x5C] = ix + 0x14;
+        }
+
+        for (ix = 0; ix < 9; ix++) {
+            if (ix == 8) {
+                disp_index = 1;
+            } else {
+                disp_index = 0;
+            }
+
+            effect_23_init(0, ix + 0x78, 0, 2, disp_index, ix, 0x70A7, ix, 0);
+            Order[ix + 0x78] = 1;
+            Order_Dir[ix + 0x78] = 4;
+            Order_Timer[ix + 0x78] = ix + 0x14;
+            effect_23_init(1, ix + 0x81, 0, 2, disp_index, ix, 0x70A7, ix, 0);
+            Order[ix + 0x81] = 1;
+            Order_Dir[ix + 0x81] = 4;
+            Order_Timer[ix + 0x81] = ix + 0x14;
+        }
+
+        Menu_Cursor_Move = 0x22;
+        effect_66_init(0x8A, 7, 2, 0, -1, -1, -0x7FFF);
+        Order[0x8A] = 1;
+        Order_Dir[0x8A] = 4;
+        Order_Timer[0x8A] = 0x14;
+        effect_66_init(0x8B, 8, 2, 0, -1, -1, -0x7FFF);
+        Order[0x8B] = 1;
+        Order_Dir[0x8B] = 4;
+        Order_Timer[0x8B] = 0x14;
+        break;
+
+    case 1:
+        Menu_Sub_case1(task_ptr);
+        break;
+
+    case 2:
+        if (FadeIn(1, 0x19, 8) != 0) {
+            task_ptr->r_no[2] += 1;
+            Suicide[3] = 0;
+        }
+
+        break;
+
+    case 3:
+        Button_Config_Sub(0);
+        Button_Exit_Check(task_ptr, 0);
+        Button_Config_Sub(1);
+        Button_Exit_Check(task_ptr, 1);
+        Save_Game_Data();
+        break;
+    }
+}
+
+void Button_Config_Sub(s16 PL_id) {
+    u16 sw = ~plsw_01[PL_id] & plsw_00[PL_id];
+    sw = Check_Menu_Lever(PL_id, 0);
+    MC_Move_Sub(sw, PL_id, 0xA, 0xFF);
+    Button_Move_Sub_LR(sw, PL_id);
+
+    if (ppwork[0].ok_dev == 0) {
+        Convert_Buff[1][0][8] = 0;
+    }
+
+    if (ppwork[1].ok_dev == 0) {
+        Convert_Buff[1][1][8] = 0;
+    }
+}
+
+void Button_Move_Sub_LR(u16 sw, s16 cursor_id) {
+    s16 max;
+
+    switch (Menu_Cursor_Y[cursor_id]) {
+    case 8:
+        max = 1;
+        break;
+
+    case 9:
+    case 10:
+        max = 0;
+        break;
+
+    default:
+        max = 11;
+        break;
+    }
+
+    if (max == 0) {
+        return;
+    }
+
+    switch (sw) {
+    case 4:
+        Convert_Buff[1][cursor_id][Menu_Cursor_Y[cursor_id]] -= 1;
+
+        if (Convert_Buff[1][cursor_id][Menu_Cursor_Y[cursor_id]] < 0) {
+            Convert_Buff[1][cursor_id][Menu_Cursor_Y[cursor_id]] = max;
+        }
+
+        if (Menu_Cursor_Y[cursor_id] == 8) {
+            if (Convert_Buff[1][cursor_id][8]) {
+                pp_vib_on(cursor_id);
+            } else {
+                pulpul_stop2(cursor_id);
+            }
+        }
+
+        SE_dir_cursor_move();
+        break;
+
+    case 8:
+        Convert_Buff[1][cursor_id][Menu_Cursor_Y[cursor_id]] += 1;
+
+        if (Convert_Buff[1][cursor_id][Menu_Cursor_Y[cursor_id]] > max) {
+            Convert_Buff[1][cursor_id][Menu_Cursor_Y[cursor_id]] = 0;
+        }
+
+        if ((Menu_Cursor_Y[cursor_id] == 8) && (Convert_Buff[1][cursor_id][Menu_Cursor_Y[cursor_id]] == 1)) {
+            pp_vib_on(cursor_id);
+        }
+
+        SE_dir_cursor_move();
+        break;
+    }
+}
+
+void Button_Exit_Check(struct _TASK* task_ptr, s16 PL_id) {
+    switch (IO_Result) {
+    case 0x200:
+    case 0x100:
+        break;
+
+    default:
+        return;
+    }
+
+    switch (task_ptr->r_no[1]) {
+    case 9:
+        if (Menu_Cursor_Y[0] == 11 || IO_Result == 0x200) {
+            SE_selected();
+            Return_Option_Mode_Sub(task_ptr);
+            Order[0x6A] = 4;
+            Order_Timer[0x6A] = 4;
+            return;
+        }
+
+        if (Menu_Cursor_Y[0] == 10) {
+            SE_selected();
+            save_w[1].Difficulty = Game_Default_Data.Difficulty;
+            save_w[1].Time_Limit = Game_Default_Data.Time_Limit;
+            save_w[1].Battle_Number[0] = Game_Default_Data.Battle_Number[0];
+            save_w[1].Battle_Number[1] = Game_Default_Data.Battle_Number[1];
+            save_w[1].Damage_Level = Game_Default_Data.Damage_Level;
+            save_w[1].GuardCheck = Game_Default_Data.GuardCheck;
+            save_w[1].AnalogStick = Game_Default_Data.AnalogStick;
+            save_w[1].Handicap = Game_Default_Data.Handicap;
+            save_w[1].Partner_Type[0] = Game_Default_Data.Partner_Type[0];
+            save_w[1].Partner_Type[1] = Game_Default_Data.Partner_Type[1];
+            Copy_Save_w();
+            return;
+        }
+
+        break;
+
+    case 10:
+        if ((Menu_Cursor_Y[PL_id] == 10) || (IO_Result == 0x200)) {
+            SE_selected();
+            Return_Option_Mode_Sub(task_ptr);
+            Order[0x6B] = 4;
+            Order_Timer[0x6B] = 4;
+            return;
+        }
+
+        if (Menu_Cursor_Y[PL_id] == 9) {
+            SE_selected();
+            Setup_IO_ConvDataDefault(PL_id);
+            Save_Game_Data();
+            return;
+        }
+
+        break;
+
+    case 13:
+        if (IO_Result == 0x200) {
+            SE_selected();
+            Return_Option_Mode_Sub(task_ptr);
+            Order[0x69] = 4;
+            Order_Timer[0x69] = 4;
+            return;
+        }
+
+        switch (Menu_Cursor_Y[0]) {
+        case 2:
+            SE_selected();
+            Return_Option_Mode_Sub(task_ptr);
+            Order[0x69] = 4;
+            Order_Timer[0x69] = 4;
+            break;
+
+        case 0:
+            SE_selected();
+            task_ptr->r_no[2] = 4;
+            task_ptr->r_no[3] = 0;
+            break;
+
+        case 1:
+            SE_selected();
+            task_ptr->r_no[2] = 5;
+            task_ptr->r_no[3] = 0;
+            break;
+        }
+
+        break;
+    }
+}
+
+void Return_Option_Mode_Sub(struct _TASK* task_ptr) {
+    Menu_Suicide[1] = 0;
+    Menu_Suicide[2] = 1;
+    task_ptr->r_no[1] = 7;
+    task_ptr->r_no[2] = 0;
+    task_ptr->r_no[3] = 0;
+    task_ptr->free[0] = 0;
+    Cursor_Y_Pos[0][2] = Menu_Cursor_Y[0];
+    Cursor_Y_Pos[1][2] = Menu_Cursor_Y[1];
+}
+
+void Screen_Adjust(struct _TASK* task_ptr) {
+    s16 char_index;
+    s16 ix;
+
+    s16 unused_s3;
+    s16 unused_s2;
+
+    X_Adjust = X_Adjust_Buff[0];
+    X_Adjust_Buff[0] = X_Adjust_Buff[1];
+    X_Adjust_Buff[1] = X_Adjust_Buff[2];
+    Y_Adjust = Y_Adjust_Buff[0];
+    Y_Adjust_Buff[0] = Y_Adjust_Buff[1];
+    Y_Adjust_Buff[1] = Y_Adjust_Buff[2];
+
+    switch (task_ptr->r_no[2]) {
+    case 0:
+        FadeOut(1, 0xFF, 8);
+        task_ptr->r_no[2] += 1;
+        task_ptr->timer = 5;
+        Menu_Common_Init();
+        Menu_Cursor_Y[0] = 0;
+        Menu_Suicide[1] = 1;
+        Menu_Suicide[2] = 0;
+        Order[0x4F] = 4;
+        Order_Timer[0x4F] = 1;
+        Order[0x4E] = 2;
+        Order_Dir[0x4E] = 2;
+        Order_Timer[0x4E] = 1;
+        effect_57_init(0x65, MENU_HEADER_SCREEN_ADJUST, 0, 0x3F, 2);
+        Order[0x65] = 1;
+        Order_Dir[0x65] = 8;
+        Order_Timer[0x65] = 1;
+
+        Convert_Buff[2][0][4] = mpp_w.language;
+
+        for (ix = 0; ix < 4; ix++) {
+            effect_63_init(ix + 0x66, 0, 2, ix, ix);
+            Order[ix + 0x66] = 1;
+            Order_Dir[ix + 0x66] = 4;
+            Order_Timer[ix + 0x66] = ix + 0x14;
+        }
+
+        effect_64_init(0x6A, 0, 2, 9, 4, 0x7047, 18, 2, 0);
+        Order[0x6A] = 1;
+        Order_Dir[0x6A] = 4;
+        Order_Timer[0x6A] = 0x18;
+
+        for (ix = 0, unused_s3 = char_index = 0xE; ix < 7; ix++, unused_s2 = char_index++) {
+            effect_61_init(0, ix + 0x50, 0, 2, char_index, ix, 0x7047);
+            Order[ix + 0x50] = 1;
+            Order_Dir[ix + 0x50] = 4;
+            Order_Timer[ix + 0x50] = ix + 0x14;
+        }
+
+        Menu_Cursor_Move = 5;
+        break;
+
+    case 1:
+        Menu_Sub_case1(task_ptr);
+        break;
+
+    case 2:
+        if (FadeIn(1, 0x19, 8) != 0) {
+            task_ptr->r_no[2] += 1;
+            Suicide[3] = 0;
+        }
+
+        break;
+
+    case 3:
+        Screen_Adjust_Sub(0);
+        Screen_Exit_Check(task_ptr, 0);
+
+        if (IO_Result == 0) {
+            Screen_Adjust_Sub(1);
+            Screen_Exit_Check(task_ptr, 0);
+        }
+
+        Save_Game_Data();
+        break;
+    }
+}
+
+void Screen_Adjust_Sub(s16 PL_id) {
+    u16 sw;
+    sw = ~plsw_01[PL_id] & plsw_00[PL_id];
+    sw = Check_Menu_Lever(PL_id, 0);
+    MC_Move_Sub(sw, 0, 6, 0xFF);
+    Screen_Move_Sub_LR(sw);
+    Convert_Buff[2][0][0] = X_Adjust_Buff[2] & 0xFF;
+    Convert_Buff[2][0][1] = Y_Adjust_Buff[2] & 0xFF;
+    Convert_Buff[2][0][2] = dspwhPack(Disp_Size_H, Disp_Size_V);
+    save_w[1].Screen_Size = dspwhPack(Disp_Size_H, Disp_Size_V);
+    Convert_Buff[2][0][4] = mpp_w.language;
+}
+
+void Screen_Exit_Check(struct _TASK* task_ptr, s16 PL_id) {
+    switch (IO_Result) {
+    case 0x200:
+    case 0x100:
+        break;
+
+    default:
+        return;
+    }
+
+    if (Menu_Cursor_Y[0] == 6 || IO_Result == 0x200) {
+        SE_selected();
+        Menu_Suicide[1] = 0;
+        Menu_Suicide[2] = 1;
+        X_Adjust = X_Adjust_Buff[2];
+        Y_Adjust = Y_Adjust_Buff[2];
+        Return_Option_Mode_Sub(task_ptr);
+
+        if (task_ptr->r_no[0] == 1) {
+            task_ptr->r_no[1] = 1;
+        } else {
+            task_ptr->r_no[1] = 7;
+            Order[0x65] = 4;
+            Order_Timer[0x65] = 4;
+        }
+
+        task_ptr->r_no[2] = 0;
+        task_ptr->r_no[3] = 0;
+        task_ptr->free[0] = 0;
+        return;
+    }
+
+    if (Menu_Cursor_Y[PL_id] == 5) {
+        SE_selected();
+        X_Adjust_Buff[2] = 0;
+        Y_Adjust_Buff[2] = 0;
+        Disp_Size_H = 100;
+        Disp_Size_V = 100;
+        mpp_w.language = Get_Default_Language();
+    }
+}
+
+u16 Sound_Cursor_Sub(s16 PL_id) {
+    u16 sw;
+    u16 ret;
+
+    sw = ~plsw_01[PL_id] & plsw_00[PL_id];
+    sw = Check_Menu_Lever(PL_id, 0);
+    ret = MC_Move_Sub(sw, 0, 5, 0xFF);
+    ret |= SD_Move_Sub_LR(sw);
+    ret &= 0x20F;
+    return ret;
+}
+
+const u8 Sound_Data_Max[3][5] = { { 0, 0, 1, 0, 66 }, { 15, 15, 1, 0, 66 }, { 15, 15, 0, 0, 0 } };
+
+u16 SD_Move_Sub_LR(u16 sw) {
+    u16 rnum;
+    s16 max;
+    s8 last_cursor;
+
+    rnum = 0;
+
+    if (Menu_Cursor_Y[0] == 3 || Menu_Cursor_Y[0] == 5) {
+        return 0;
+    }
+
+    last_cursor = Convert_Buff[3][1][Menu_Cursor_Y[0]];
+
+    switch (sw) {
+    case 4:
+        max = Sound_Data_Max[0][Menu_Cursor_Y[0]];
+
+        while (1) {
+            Convert_Buff[3][1][Menu_Cursor_Y[0]] -= 1;
+
+            if (Convert_Buff[3][1][Menu_Cursor_Y[0]] < 0) {
+                Convert_Buff[3][1][Menu_Cursor_Y[0]] = max;
+            }
+
+            if ((Menu_Cursor_Y[0] != 4) || (bgmSkipCheck(Convert_Buff[3][1][4] + 1) == 0)) {
+                break;
+            }
+        }
+
+        if (last_cursor != Convert_Buff[3][1][Menu_Cursor_Y[0]]) {
+            rnum = 4;
+        }
+
+        break;
+
+    case 8:
+        max = Sound_Data_Max[1][Menu_Cursor_Y[0]];
+
+        while (1) {
+            Convert_Buff[3][1][Menu_Cursor_Y[0]] += 1;
+
+            if (Convert_Buff[3][1][Menu_Cursor_Y[0]] > max) {
+                Convert_Buff[3][1][Menu_Cursor_Y[0]] = Sound_Data_Max[2][Menu_Cursor_Y[0]];
+            }
+
+            if ((Menu_Cursor_Y[0] != 4) || (bgmSkipCheck(Convert_Buff[3][1][4] + 1) == 0)) {
+                break;
+            }
+        }
+
+        if (last_cursor != Convert_Buff[3][1][Menu_Cursor_Y[0]]) {
+            rnum = 8;
+        }
+
+        break;
+    }
+
+    if (rnum) {
+        SE_dir_cursor_move();
+    }
+
+    return rnum;
+}
+
+s32 Setup_Final_Cursor_Pos(s8 cursor_x, s16 dir) {
+    s16 ix;
+    s16 check_x[2];
+    s16 next_dir;
+
+    if (cursor_x == -1) {
+        cursor_x = 0;
+    }
+
+    if (vm_w.Connect[cursor_x]) {
+        return cursor_x;
+    }
+
+    check_x[0] = cursor_x ^ 1;
+
+    if (vm_w.Connect[check_x[0]]) {
+        return check_x[0];
+    }
+
+    if (dir == 4) {
+        next_dir = -2;
+    } else {
+        next_dir = 2;
+    }
+
+    check_x[0] = cursor_x;
+
+    for (ix = 0; ix < 4; ix++) {
+        check_x[0] += next_dir;
+
+        if (check_x[0] < 0) {
+            if (IO_Result == 0) {
+                check_x[0] += 8;
+            } else {
+                return Menu_Cursor_X[1];
+            }
+        }
+
+        if (check_x[0] > 7) {
+            if (IO_Result == 0) {
+                check_x[0] -= 8;
+            } else {
+                return Menu_Cursor_X[1];
+            }
+        }
+
+        if (vm_w.Connect[check_x[0]]) {
+            return check_x[0];
+        }
+
+        check_x[1] = check_x[0] ^ 1;
+
+        if (vm_w.Connect[check_x[1]]) {
+            return check_x[1];
+        }
+    }
+
+    return -1;
+}
+
+static bool should_skip_menu_cursor(s16 cursor_id, s16 cansel_menu) {
+    return (cansel_menu == Menu_Cursor_Y[cursor_id]) && (Connect_Status == 0);
+}
+
+u16 MC_Move_Sub(u16 sw, s16 cursor_id, s16 menu_max, s16 cansel_menu) {
+    if (Menu_Cursor_Move > 0) {
+        return 0;
+    }
+
+    switch (sw) {
+    case SWK_UP:
+        Menu_Cursor_Y[cursor_id] -= 1;
+
+        if (Menu_Cursor_Y[cursor_id] < 0) {
+            Menu_Cursor_Y[cursor_id] = menu_max;
+        }
+
+        if (should_skip_menu_cursor(cursor_id, cansel_menu)) {
+            Menu_Cursor_Y[cursor_id] -= 1;
+        }
+
+        SE_cursor_move();
+        return IO_Result = SWK_UP;
+
+    case SWK_DOWN:
+        Menu_Cursor_Y[cursor_id] += 1;
+
+        if (Menu_Cursor_Y[cursor_id] > menu_max) {
+            Menu_Cursor_Y[cursor_id] = 0;
+        }
+
+        if (should_skip_menu_cursor(cursor_id, cansel_menu)) {
+            Menu_Cursor_Y[cursor_id] += 1;
+        }
+
+        SE_cursor_move();
+        return IO_Result = SWK_DOWN;
+
+    case SWK_WEST:
+        return IO_Result = SWK_WEST;
+
+    case SWK_SOUTH:
+        return IO_Result = SWK_SOUTH;
+
+    case SWK_EAST:
+        return IO_Result = SWK_EAST;
+
+    case SWK_RIGHT_TRIGGER:
+        return IO_Result = SWK_RIGHT_TRIGGER;
+
+    case SWK_START:
+        return IO_Result = SWK_START;
+
+    default:
+        return IO_Result = 0;
+
+    case SWK_NORTH:
+        return IO_Result = SWK_NORTH;
+
+    case SWK_RIGHT_SHOULDER:
+        return IO_Result = SWK_RIGHT_SHOULDER;
+
+    case SWK_LEFT_SHOULDER:
+        return IO_Result = SWK_LEFT_SHOULDER;
+
+    case SWK_LEFT_TRIGGER:
+        return IO_Result = SWK_LEFT_TRIGGER;
+    }
+}
+
+s32 Exit_Sub(struct _TASK* task_ptr, s16 cursor_ix, s16 next_routine) {
+    switch (task_ptr->free[0]) {
+    case 0:
+        task_ptr->free[0] += 1;
+        FadeInit();
+        /* fallthrough */
+
+    case 1:
+        if (!FadeOut(1, 25, 8)) {
+            return 0;
+        }
+
+        task_ptr->r_no[1] = next_routine;
+        task_ptr->r_no[2] = 0;
+        task_ptr->r_no[3] = 0;
+        task_ptr->free[0] = 0;
+        Cursor_Y_Pos[0][cursor_ix] = Menu_Cursor_Y[0];
+        Cursor_Y_Pos[1][cursor_ix] = Menu_Cursor_Y[1];
+        pulpul_stop();
+        return 1;
+
+    default:
+        return 0;
+    }
+}
