+++ src/sf33rd/Source/Game/menu/screen_adjust.c
@@ -1,0 +1,130 @@
+#include "sf33rd/Source/Game/menu/menu.h"
+#include "common.h"
+#include "main.h"
+#include "sf33rd/Source/Game/engine/workuser.h"
+#include "sf33rd/Source/Game/menu/dir_data.h"
+#include "sf33rd/Source/Game/sound/se.h"
+#include "sf33rd/Source/Game/sound/sound3rd.h"
+#include "sf33rd/Source/Game/system/work_sys.h"
+#include "structs.h"
+
+static s16 adjust_screen_size(u16 sw) {
+    s16 flag = 0;
+
+    if (sw == 4) {
+        switch (Menu_Cursor_Y[0]) {
+        case 0:
+            X_Adjust_Buff[2] -= 2;
+
+            if (X_Adjust_Buff[2] < -10) {
+                X_Adjust_Buff[2] = -10;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 1:
+            Y_Adjust_Buff[2] -= 2;
+
+            if (Y_Adjust_Buff[2] < -10) {
+                Y_Adjust_Buff[2] = -10;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 2:
+            Disp_Size_H -= 2;
+
+            if (Disp_Size_H < 94) {
+                Disp_Size_H = 94;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 3:
+            Disp_Size_V -= 2;
+
+            if (Disp_Size_V < 94) {
+                Disp_Size_V = 94;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 4:
+            mpp_w.language = Language_Toggle(mpp_w.language);
+            flag = 1;
+            break;
+        }
+    } else if (sw == 8) {
+        switch (Menu_Cursor_Y[0]) {
+        case 0:
+            X_Adjust_Buff[2] += 2;
+
+            if (X_Adjust_Buff[2] > 10) {
+                X_Adjust_Buff[2] = 10;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 1:
+            Y_Adjust_Buff[2] += 2;
+
+            if (Y_Adjust_Buff[2] > 10) {
+                Y_Adjust_Buff[2] = 10;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 2:
+            Disp_Size_H += 2;
+
+            if (Disp_Size_H > 100) {
+                Disp_Size_H = 100;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 3:
+            Disp_Size_V += 2;
+
+            if (Disp_Size_V > 100) {
+                Disp_Size_V = 100;
+            } else {
+                flag = 1;
+            }
+
+            break;
+
+        case 4:
+            mpp_w.language = Language_Toggle(mpp_w.language);
+            flag = 1;
+            break;
+        }
+    }
+
+    return flag;
+}
+
+void Screen_Move_Sub_LR(u16 sw) {
+    s16 flag = adjust_screen_size(sw);
+
+    if (flag) {
+        SE_dir_cursor_move();
+    }
+
+    X_Adjust = X_Adjust_Buff[0] = X_Adjust_Buff[1] = X_Adjust_Buff[2];
+    Y_Adjust = Y_Adjust_Buff[0] = Y_Adjust_Buff[1] = Y_Adjust_Buff[2];
+}
